backend/requirements.txt
fast_bitrix24==1.8.5
Flask==2.2.3
python-dotenv==1.1.0
fastapi
uvicorn
aiohttp
